move test lua files to test folder
