+++ justfile
@@ -7,3 +7,9 @@ lint:
 
 dev:
     docker-compose up --build
+
+export DOCKER_CLI_HINTS := "false"
+
+oc-test:
+    docker build -t oc tests/test_opencomputers
+    docker run -it --rm --network="host" oc
