--- opencomputers/instructions.txt
@@ -1,15 +1,0 @@
-docker build -t oc .
-docker run -it --network="host" oc   
-lua
-dofile("boot2.lua")
-co=boot_machine()
-coroutine.resume(co)
-coroutine.resume(co)
-coroutine.resume(co)
-coroutine.resume(co)
-
-========
-
-docker build -t oc .
-docker run -it --network="host" oc   
-lua test.lua
