+++ tests/test_opencomputers/ocemu.cfg
